update migration file

+++ db/migrate/20160820004638_add_attachment_attachment_to_carrier_types.rb
@@ -3,7 +3,7 @@ class AddAttachmentAttachmentToCarrierTypes < ActiveRecord::Migration[4.2]
     change_table :carrier_types do |t|
       t.string :attachment_content_type
       t.string :attachment_file_name
-      t.bigint :attachment_file_size
+      t.integer :attachment_file_size
       t.datetime :attachment_updated_at
     end
   end
